third_party/WebKit/LayoutTests/inspector/console/console-message-format-expected.txt

CONSOLE MESSAGE: line 9: Message format number %i, %d and %f
CONSOLE MESSAGE: line 10: Message %s for %s
CONSOLE MESSAGE: line 11: Object %o
CONSOLE MESSAGE: line 12: Array %o
CONSOLE MESSAGE: line 13: Object as object: %O
CONSOLE MESSAGE: line 14: Array as object: %O
CONSOLE MESSAGE: line 15: Floating as integers: %d %i
CONSOLE MESSAGE: line 16: Floating as is: %f
CONSOLE MESSAGE: line 17: Non-numbers as numbers: %d %i %f
CONSOLE MESSAGE: line 18: String as is: %s
CONSOLE MESSAGE: line 19: Object as string: %s
Tests that console logging uses proper message formatting.

console-message-format.html:9 Message format number 1, 2 and 3.5
console-message-format.html:10 Message format for string
console-message-format.html:11 Object {foo: "bar"}
console-message-format.html:12 Array (2) ["foo", "bar"]
console-message-format.html:13 Object as object: Object
console-message-format.html:14 Array as object: Array(2)
console-message-format.html:15 Floating as integers: 42 42
console-message-format.html:16 Floating as is: 42.5
console-message-format.html:17 Non-numbers as numbers: NaN NaN NaN
console-message-format.html:18 String as is: string
console-message-format.html:19 Object as string: #document
